fix pidXOR type

+++ Source/Core/Gen8/Generators/EggGenerator8.cpp
@@ -188,7 +188,7 @@ std::vector<EggGeneratorState> EggGenerator8::generate(u64 seed0, u64 seed1) con
                     break;
                 }
             }
-            u8 pidXOR = psv ^ tsv;
+            u16 pidXOR = psv ^ tsv;
             u8 shiny = pidXOR < 16 ? pidXOR == 0 ? 2 : 1 : 0;
 
             // Ball handling check
